requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==24.3.0
certifi==2024.12.14
deprecation==2.1.0
dj-database-url==2.3.0
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
frozenlist==1.5.0
gotrue==2.11.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
multidict==6.1.0
packaging==24.2
postgrest==0.19.1
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
realtime==2.1.0
rest-framework-simplejwt==0.0.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
storage3==0.11.0
StrEnum==0.4.15
supabase==2.11.0
supafunc==0.9.0
typing_extensions==4.12.2
websockets==13.1
yarl==1.18.3
whitenoise
gunicorn